requirements.txt
httpx==0.28.1
loguru==0.7.3
pydantic==2.11.3
pydantic_core==2.33.1
pydantic_settings==2.9.1
python-telegram-bot==22.0
pytest==8.3.5
pytest-asyncio==0.26.0